move cloudwatch backdoor API to asf provider
